fix(web): keep pg out of automation knowledge client bundle
The knowledge files panel imported limits from drizzle-backed modules and accept extensions from unpdf/mammoth extractors, so Next still bundled pg into Client Component Browser after the public chat split. Move client-safe constants and types to their own module and mark the store and extractors server-only.
